// Frontend/src/app/generated-api/model/createGroupRequest.ts
/**
 * Device Manager API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Request body for creating a new group with its initial admin user
 */
export interface CreateGroupRequest { 
    /**
     * Name of the group
     */
    groupName: string;
    /**
     * Access code used to enter the group
     */
    accessCode: string;
    /**
     * Name of the initial admin user
     */
    adminUserName: string;
    /**
     * Password of the initial admin user
     */
    adminPassword: string;
}