// Frontend/src/app/generated-api/model/validateGroupAccessRequest.ts
/**
 * Device Manager API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Request body for validating a group\'s access code
 */
export interface ValidateGroupAccessRequest { 
    /**
     * Access code of the group
     */
    accessCode: string;
}